Resize and rotate downloaded images
Images are resized to the configured photo frame resolution to save space
on the SD card and also to make it quicker to browse through the images
in case the photo frame is slow to render big files.

Images are rotated according to the EXIF rotation information inside the
images. This helps with photo frames which don't read the EXIF data.

Resizing and rotating images is slow and requires more memory than the
zsun has; let's use a swap file on the SD card to avoid OOM problems.